prevent finding identical routes multiple times
If a zone contained multiple stops, and transfers between the stops were possible, a route would be found multiple times, once ending at the actual stop, and additional times ending with a transfer to one of the other stops in the zone.

The same could also happen at the start of the route, with first a transfer from a stop to another one where the actual trip starts.

Closes #13

Solves part of #15 with double routes LS-GE.