Fix a bug where wrong access/egress time might have been used
If two zones connect to the same stop, but with different access times, it occurred that the second zone overwrote the access times for the first zone. This was due to the fact that a single FoundRoute instance was used for both OD-relations (the one ending in the first zone as well as the one ending in the second zone).

The code now differentiates between a Stop2StopRoute (the route returned from SwissRailRaptor describing the pure route between two stops, but independent of any access/egress). The existing FoundRoute was adapted to use a Stop2StopRoute, but add the zone access information in addition. This way, multiple FoundRoutes can be created for a single Stop2StopRoute, solving the problem with overwriting data. This also helps to get rid of the workaround in FoundRoute where the demand and adaptionTime had to be stored in a Map for different zones, as now each FoundRoute is for a single (destination) zone.

Fixes #15